Modify project to show more details on instances, etc
This commit adds an ANCOR version number to the title page. This commit
adds an alert function that will be given a msg containing instance
information from the view. Eventually this will be used to generate a
detailed modal view of each instance. Other information was added to
readme for future planning.