Add modal view to instance details
This commit adds a modal view for each instance within the instance
details table. For now it will just display the simple information
provided in instances.json, but in the future it will do an HTTP GET
call on the specific instance ID to show more information in the modal
about the given instance.