Modify how add new instance button works and route reload
This commit changes the add new instance button to be a group dropdown
that will contain all available roles from ancor. This commit also adds
a reload feature that will reload the page when someone replaces or
deletes and instance. This will trigger a new HTTP GET call on all
instances to reflect the change on the dashboard.