Add ACE template button dropdown
This commit adds the ACE template button dropdown. If a user wants to
add a goal or role template, they are able to select either 'goal' or
'role' from the dropdown above the editor, and depending on which one
they select it will insert a sample template block for them to fill out.