Hook up ancor-dashboard to live ancor
This commit hooks up ancor-dashboard to a live version of ancor. Changes
have been made to fit within new API so data is properly shown. ancor ip
address is stored in the root scope so all controllers may access it if
an http call is required. example yaml conf is modified to fit updated
syntax in snake case.