Add delete button to env, add deploy functionality
This commit adds the delete functionality to the env endpoint. It makes
an http.delete to remove the environment. The deploy functionality makes
the correct http post call for planning (with the correct data type
header) and then the correct http put call for commiting.

There were some errors once I clicked deploy in the js console that
might be an issue. Will be debugged later.